requirements.txt
fastapi
uvicorn
azure.monitor.opentelemetry